chore: initial cleanup
