--- app/(personal)/components/ThemeToggle.tsx
@@ -1,60 +1,0 @@
-'use client'
-import { useTheme } from 'next-themes'
-import { useEffect, useState } from 'react'
-
-export function ThemeToggle() {
-  const [mounted, setMounted] = useState(false)
-  const { theme, setTheme } = useTheme()
-
-  useEffect(() => {
-    setMounted(true)
-  }, [])
-  if (!mounted) return null
-
-  return (
-    <div className="flex items-center justify-center">
-      <button
-        type="button"
-        aria-label="Theme toggle button"
-        className="inline-flex items-center rounded p-2 font-medium shadow-sm duration-500 hover:bg-gray-100 dark:hover:bg-gray-600"
-        onClick={() => setTheme(theme === 'dark' ? 'light' : 'dark')}
-      >
-        {theme === 'light' ? moonIcon : sunIcon}
-      </button>
-    </div>
-  )
-}
-
-const moonIcon = (
-  <svg
-    className="h-4 w-4"
-    fill="transparent"
-    stroke="black"
-    viewBox="0 0 24 24"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="2"
-      d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z"
-    ></path>
-  </svg>
-)
-
-const sunIcon = (
-  <svg
-    className="h-4 w-4"
-    fill="transparent"
-    stroke="white"
-    viewBox="0 0 24 24"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="2"
-      d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z"
-    ></path>
-  </svg>
-)
