--- app/(personal)/components/index.tsx
@@ -1,5 +1,0 @@
-export { Footer } from './Footer'
-export { Header } from './Header'
-export { Navbar } from './Navbar'
-export { Project } from './Project'
-export { ProjectListItem } from './ProjectListItem'
