--- app/(personal)/providers.tsx
@@ -1,7 +1,0 @@
-'use client'
-
-import { ThemeProvider } from 'next-themes'
-
-export function Providers({ children }: { children: React.ReactNode }) {
-  return <ThemeProvider attribute="class">{children}</ThemeProvider>
-}
