+++ app/layout.tsx
@@ -1,6 +1,23 @@
 import 'tailwindcss/tailwind.css'
 
-import { Providers } from './(personal)/providers'
+import { IBM_Plex_Mono, Inter, PT_Serif } from '@next/font/google'
+
+const serif = PT_Serif({
+  variable: '--font-serif',
+  style: ['normal', 'italic'],
+  subsets: ['latin'],
+  weight: ['400', '700'],
+})
+const sans = Inter({
+  variable: '--font-sans',
+  subsets: ['latin'],
+  weight: ['500', '700', '800'],
+})
+const mono = IBM_Plex_Mono({
+  variable: '--font-mono',
+  subsets: ['latin'],
+  weight: ['500', '700'],
+})
 
 export default async function RootLayout({
   children,
@@ -8,10 +25,11 @@ export default async function RootLayout({
   children: React.ReactNode
 }) {
   return (
-    <html lang="en">
-      <body className="bg-white text-black dark:bg-black dark:text-white">
-        <Providers>{children}</Providers>
-      </body>
+    <html
+      lang="en"
+      className={`${mono.variable} ${sans.variable} ${serif.variable}`}
+    >
+      <body>{children}</body>
     </html>
   )
 }
