+++ components/Header.tsx
@@ -1,5 +1,6 @@
 import { PortableText } from '@portabletext/react'
 import React from 'react'
+
 interface HeaderProps {
   title?: string
   description?: any[]
