+++ components/ProjectListItem.tsx
@@ -1,7 +1,8 @@
 import { PortableText } from '@portabletext/react'
-import { ShowcaseProjects } from 'app/(personal)/queries'
 import { urlForImage } from 'lib/sanity.image'
 import Image from 'next/image'
+
+import type { ShowcaseProjects } from '../types'
 
 interface ProjectProps {
   project?: ShowcaseProjects
