+++ components/TimelineSection.tsx
@@ -1,6 +1,5 @@
-import { MilestoneItem } from 'app/(personal)/about/queries'
-
-import { TimelineItem } from './timeline-item'
+import type { MilestoneItem } from '../types'
+import { TimelineItem } from './TimelineItem'
 
 interface TimelineItem {
   title: string
@@ -10,13 +9,13 @@ interface TimelineItem {
 export function TimelineSection({ timelines }: { timelines: TimelineItem[] }) {
   return (
     <div className="flex gap-2 pt-16">
-      {timelines.map((timeline, key) => {
+      {timelines?.map((timeline, key) => {
         const { title, milestones } = timeline
         return (
           <div className="max-w-[50%]" key={key}>
             <div className="text-l pb-5 font-bold">{title}</div>
 
-            {milestones.map((experience, index) => (
+            {milestones?.map((experience, index) => (
               <div key={index}>
                 <TimelineItem
                   milestone={experience}
