+++ package.json
@@ -14,7 +14,7 @@
     "singleQuote": true
   },
   "dependencies": {
-    "@next/font": "^13.0.5",
+    "@next/font": "^13.0.6",
     "@portabletext/react": "^2.0.0",
     "@sanity/demo": "^1.0.1",
     "@sanity/image-url": "^1.0.1",
@@ -26,8 +26,7 @@
     "intl-segmenter-polyfill": "^0.4.4",
     "next": "^13.0.6",
     "next-google-fonts": "^2.2.0",
-    "next-sanity": "^3.0.5",
-    "next-themes": "^0.2.1",
+    "next-sanity": "^3.1.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-is": "^18.2.0",
@@ -46,7 +45,7 @@
     "postcss": "^8.4.19",
     "prettier": "^2.8.0",
     "prettier-plugin-packagejson": "^2.3.0",
-    "prettier-plugin-tailwindcss": "^0.1.13",
+    "prettier-plugin-tailwindcss": "^0.2.0",
     "tailwindcss": "^3.2.4",
     "typescript": "^4.9.3"
   }
