+++ schemas/singletons/home.ts
@@ -44,10 +44,10 @@ export default defineType({
         'These are the projects that will appear first on your landing page.',
       type: 'array',
       of: [
-        {
+        defineArrayMember({
           type: 'reference',
           to: [{ type: 'project' }],
-        },
+        }),
       ],
     }),
   ],
